ARTEMIS-1016 Fix Block Flush on Replication Manager
When using the same connection for protocol packets and for cluster
connections it is possible that contention can occur on the remoting
connection.  This happens as when a packet arrives on the remoting
connection a transfer lock is aquired locking the connection.

When replicating the replication manager will attempt to send the
replicate packet down the same connection.  For this it needs to check
that the connection is writable (i.e. buffer is not full).  If it is not
writable it will wait, until it becomes writable or times out.  However,
since the transfer lock is still aquired no flushes are able to take
place on the connection.  Meaning that if the connection is not writable
at the moment the transfer lock is originally acquired, the replication
manager will hang until it's cluster connection times out.

This patch changes the logic so that if the ReplicationManager thread
holds the transfer lock it will flush.  This will cause the
RemotingConnection to become writable again.